requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
autoflake==2.3.1
Automat==22.10.0
bandit==1.7.8
betterpath==0.2.2
billiard==4.2.0
black==24.4.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-fsm==3.0.0
django-log-request-id==2.1.0
django-service-objects @ git+https://github.com/dvkurbatov/django-service-objects@a0c4b434254fb3ca9d09ee0a33ca611926e4a602
django-split-settings==1.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
filelock==3.13.4
flake8==7.0.0
gunicorn==22.0.0
hyperlink==21.0.0
identify==2.5.35
idna==3.7
incremental==22.10.0
inflection==0.5.1
isort==5.13.2
kombu==5.3.7
l==0.11.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.0.8
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pbr==6.0.0
pillow==10.3.0
platformdirs==4.2.0
pre-commit==3.7.0
pre-commit-hooks==4.6.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
pyupgrade==3.15.2
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.5.0
stevedore==5.2.0
tokenize-rt==5.2.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vcversioner==2.16.0.0
vine==5.1.0
virtualenv==20.25.2
wcwidth==0.2.13
yesqa==1.5.0
zope.interface==6.3
